requirements.txt
streamlit
torch
torchvision
transformers
facenet-pytorch
pillow
imageio
numpy
scipy
